added changes

+++ picmaker.c
@@ -12,7 +12,7 @@ int main(){
     printf("%s\n", strerror(errno));
   }
   fprintf(fp, "P3\n500 500\n255\n"); //header
-  int i, j;
+  double i, j;
   double frequency = 0.01;
   int width = 25;
   int center = 230;
@@ -20,8 +20,8 @@ int main(){
     for (j = 0; j < 500; j++){
       double shift = (j / 25 + 1) * 0.75;
       int red = sin(frequency * i - shift) * width + center;
-      int green = sin(frequency * i + 2 - shift) * width + center;
-      int blue = sin(frequency * i + 4 - shift) * width + center;
+      int green = sin(frequency * i + 2.0 - shift) * width + center;
+      int blue = sin(frequency * i + 4.0 - shift) * width + center;
       fprintf(fp, "%d %d %d\n", red, green, blue);
     }
   }
